# docs/regression-baseline/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Base64 Convertor API (Baseline)",
    "version": "0.0.1-SNAPSHOT",
    "description": "Hand-authored OpenAPI baseline. No OpenAPI/Swagger generator (e.g. springdoc-openapi) is present in pom.xml and no /v3/api-docs endpoint exists in this application, so this document was manually authored from the controller source and live baseline captures in docs/regression-baseline/, rather than exported from a running generator. Treat this as a snapshot for regression comparison, not a live contract."
  },
  "servers": [{ "url": "http://localhost:8080" }],
  "paths": {
...
    "/api/files/convert": {
      "post": {
        "summary": "Batch URL to Base64 conversion",
        "requestBody": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/FileConvertRequest" } } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/FileConvertResponse" } } } } }, "400": { "description": "Bad Request" } }
      }
    },
    "/api/files/convert/single": {
      "post": {
        "summary": "Single URL to Base64 conversion",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileConvertRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileConvertResponse" } } } } }
      }
    },
    "/api/files/status/{processingId}": {
      "get": {
        "summary": "Poll async conversion job status",
        "parameters": [{"name":"processingId","in":"path","required":true,"schema":{"type":"string"}}],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileConvertResponse" } } } }, "404": { "description": "Not Found" } }
      }
    },
...
    "/api/files/save-decoded": {
      "post": {
        "summary": "Decode Base64 to binary file + metadata",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Base64SaveRequest" } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Bad Request" }, "500": { "description": "Internal Server Error" } }
      }
    },
    "/api/files/download-decoded/{fileName}": {
      "get": {
        "summary": "Download decoded binary file",
        "parameters": [{"name":"fileName","in":"path","required":true,"schema":{"type":"string"}}],
        "responses": { "200": { "description": "OK", "content": { "application/octet-stream": {} } }, "403": { "description": "Forbidden" }, "404": { "description": "Not Found" } }
      }
    },
    "/api/files/metadata/{fileName}": {
      "get": {
        "summary": "Fetch metadata JSON for decoded file",
        "parameters": [{"name":"fileName","in":"path","required":true,"schema":{"type":"string"}}],
        "responses": { "200": { "description": "OK" }, "403": { "description": "Forbidden" }, "404": { "description": "Not Found" } }
      }
    },
...
    "/api/files/download/{fileName}": {
      "get": {
        "summary": "Download .b64 text file as attachment",
        "parameters": [{"name":"fileName","in":"path","required":true,"schema":{"type":"string"}}],
        "responses": { "200": { "description": "OK", "content": { "text/plain": {} } }, "403": { "description": "Forbidden" }, "404": { "description": "Not Found" } }
      }
    },
    "/api/files/content/{fileName}": {
      "get": {
        "summary": "Return .b64 content inline",
        "parameters": [{"name":"fileName","in":"path","required":true,"schema":{"type":"string"}}],
        "responses": { "200": { "description": "OK", "content": { "text/plain": {} } }, "403": { "description": "Forbidden" }, "404": { "description": "Not Found" } }
      }
    },
    "/api/files/{fileName}": {
      "delete": {
        "summary": "Delete a .b64 file",
        "parameters": [{"name":"fileName","in":"path","required":true,"schema":{"type":"string"}}],
        "responses": { "200": { "description": "OK" }, "403": { "description": "Forbidden" }, "404": { "description": "Not Found" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FileConvertRequest": { "type": "object", "properties": { "url": {"type":"string"}, "fileName": {"type":"string"}, "caption": {"type":"string"}, "mimeType": {"type":"string"}, "type": {"type":"string"} }, "required": ["url","mimeType","type"] },
      "FileConvertResponse": { "type": "object", "properties": { "processingId": {"type":"string","nullable":true}, "fileName": {"type":"string"}, "mimeType": {"type":"string"}, "type": {"type":"string"}, "base64Data": {"type":"string","nullable":true}, "success": {"type":"boolean"}, "message": {"type":"string"} } },
...
      "Base64SaveRequest": { "type": "object", "properties": { "base64Content": {"type":"string"}, "fileName": {"type":"string"} }, "required": ["base64Content","fileName"] }
    }
  }
}